# docs/openapi.yaml
openapi: 3.0.3
info:
  title: 2Bee API
  description: |
    API specification for the 2Bee shared finance application.
    Profile & Settings stubs live under /api (Dana Hanin).
    Dashboard & Analytics endpoints are at the server root (Mishelle — PR1).
    Remaining API groups will be added by each domain owner as per plan.md.
  version: 0.1.0-stubs
servers:
  - url: http://localhost:4000/api
    description: Profile API (path prefix /api)
  - url: http://localhost:4000
    description: Root API (dashboard, hive, expenses, auth)

# ---------------------------------------------------------------------------
# Security
# ---------------------------------------------------------------------------
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: |
        Bearer access token (opaque session token) issued by POST /auth/login.
        Include as: Authorization: Bearer <token>

  # ---------------------------------------------------------------------------
  # Reusable schemas
  # ---------------------------------------------------------------------------
  schemas:
...
    AuthUser:
      type: object
      properties:
        id:
          type: string
          example: user_demo_1
        firstName:
          type: string
          example: Demo
        lastName:
          type: string
          example: User
        email:
          type: string
          format: email
          example: demo@2bee.app
        pairId:
          type: string
          nullable: true
          example: null
        hiveId:
          type: string
          nullable: true
          example: hive_demo_1
        createdAt:
          type: string
          format: date-time
          example: 2024-06-01T09:00:00.000Z
      required:
        - id
        - firstName
        - lastName
        - email
        - createdAt
    RegisterRequest:
      type: object
      required:
        - firstName
        - lastName
        - email
        - password
      properties:
        firstName:
          type: string
          minLength: 2
          maxLength: 50
          example: Demo
        lastName:
          type: string
          minLength: 2
          maxLength: 50
          example: User
        email:
          type: string
          format: email
          example: demo@2bee.app
        password:
          type: string
          format: password
          minLength: 8
          example: hunter22
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: demo@2bee.app
        password:
          type: string
          format: password
          example: hunter22
    RegisterResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/AuthUser'
    LoginResponse:
      type: object
      properties:
        accessToken:
          type: string
          example: 51e4b5830a2743aabf98c5404f4f894c
        expiresAt:
          type: string
          format: date-time
          example: 2024-06-01T09:15:00.000Z
        user:
          $ref: '#/components/schemas/AuthUser'
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: UNAUTHORIZED
            message:
              type: string
              example: Missing or invalid authorization token.
            details:
              type: object
              nullable: true
              additionalProperties:
                type: string
...
security:
  - bearerAuth: []

# ---------------------------------------------------------------------------
# Paths
# ---------------------------------------------------------------------------
paths:
...
  /auth/register:
    post:
      summary: Register a new user account
      description: Creates an account with hashed password storage and returns the sanitized user profile.
      operationId: registerUser
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        '400':
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Email already exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/login:
    post:
      summary: Login with email and password
      description: Verifies credentials with bcrypt and issues a short-lived access token.
      operationId: loginUser
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Credentials verified.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid credentials.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/logout:
    post:
      summary: Logout and invalidate the current access token
      description: Accepts the Bearer token and removes the associated session record.
      operationId: logoutUser
      tags:
        - Authentication
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Session invalidated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Missing token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Token invalid or expired.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # --------------------------------------------------------------------------
  # Dashboard & Analytics (Mishelle — PR1)
  # --------------------------------------------------------------------------
...